refactor(core): split up interface files in render3 (#21433)
PR Close #21433
